Break out `CrateRustdoc{,JsonTypes}`/`RustdocTheme` into their own `rustdoc_tests` module
